Rewrite bootstrapping
Initialize the app in the angular `run` block. This gives us more flexibility how and when to load documents. Also we have access to dependency injection so that the bootstrapping no longer uses a different way of loading the document from the API. The API mock is replaced with a localApi document_service that can be switched out easily by a document_service which accesses the watson API.