rewrite choropleth directive for projections
so far the choropleth directive used the albersUsa projection which
of course doesn't work for anything but the USA. Now the default
projection is mercator and the directive takes another projection
function as a parameter. The viewBox is now calculated by the bounds
of the given geometry rather than the bbox attribute of the svg node.