refactor snippet data update strategy
so far updates to the snippet data were caught by setting and
listening to the data member lastChangeTime. This is cumbersome since
every time data changes this member has to be updated manually.
The livingdocs-engine was changed so that setting the data is always
done by passing a hash. Setting of all data in this hash is atomic and
only after all data was set, a callback will be fired to which the UI
can listen.
The UI listens to the data changed callback from the engine in every
instance of a choropleth and decides if the update is for the choropleth
in question (using the snippets id). Setting of lastChangeTime is
not necessary anymore.