move all scope manipulation to controller
The choropleth map model did a lot of scope manipulation. This is though
the responsibility of the controller, so I moved everything there.
ChoroplethMap#wasInserted is deprecated since it is controller 
constructor code. We need to do the same for the map model then
we can get rid of the model interface altogether.