fix unintuitive behavior in max value deduction
the max value was so far taken over all value properties of a data
set even if they had no visual representation. This makes little
sense and is confusing. Now the max value is only considered for the
value properties that are actually visible on the map, i.e. have
a mapping. This though also requires a redraw when the mapping
properties change.