simplify legend to use only classes for coloring
puts an svg with a rect in every li element of the legend. that way we
have the best of both worlds: css can be used to make the list
responsive, e.g. make the legend vertical on small devices and render
the text to the right. by using a rect we can use the color brewer css
file and don't need to assign colors with javascript. this makes
changing a color scheme much easier: just switch a class on the top
directives element.