remove autosave pause while in history modal
since we are working only on a copy of the snippet model it is not
necessary to disable autosave since this copy is in no snippet tree and
thus never sends a change event. also with the current interface it
might be a good idea not to send a special save request but just use
the regular autosave once the merge is performed.